youtube-live/requirements.txt
youtube-dl
pytchat
pafy
pandas
python-dotenv
fastapi
kafka-python
uvicorn